Initial commit: git, IdeaVim, Vim, VS Code
Let's begin!  This new set of dotfiles is maintained using GNU Stow,
which manages symlinks for package deployments.  In theory, it should
allow for a new system to be bootstrapped with only a small number of
commands; accordingly, I have migrated my Vim configuration to use git
submodules rather than a package manager to allow for "git clone" to be
run only once in a new deployment.